chore(e2e): move 2025 tests
