+++ e2e/src/test/java/dev/tachyonmcp/e2e/mcp/TypedToolRegistrationTest.java
@@ -10,7 +10,6 @@ import dev.tachyonmcp.testkit.McpTestServers;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
-
 import kotlinx.serialization.SerialName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
