+++ e2e/src/test/java/dev/tachyonmcp/e2e/mcp/v2025_11_25/LoggingTest.java
@@ -1,11 +1,10 @@
 /* Copyright (c) 2026 Konstantin Pavlov/IT Staff and contributors. */
-package dev.tachyonmcp.e2e.mcp;
+package dev.tachyonmcp.e2e.mcp.v2025_11_25;
 
 import static java.time.Duration.ofSeconds;
 import static net.javacrumbs.jsonunit.assertj.JsonAssertions.assertThatJson;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import dev.tachyonmcp.e2e.mcp.v2025_11_25.AbstractStatefulMcpE2eTest;
 import dev.tachyonmcp.testkit.SseFrame;
 import java.util.Map;
 import java.util.stream.Stream;
