+++ e2e/src/test/java/dev/tachyonmcp/e2e/mcp/v2025_11_25/StatefulMcpSdkTest.java
@@ -1,7 +1,7 @@
 /* Copyright (c) 2026 Konstantin Pavlov/IT Staff and contributors. */
-package dev.tachyonmcp.e2e.mcp;
+package dev.tachyonmcp.e2e.mcp.v2025_11_25;
 
-import dev.tachyonmcp.e2e.mcp.v2025_11_25.AbstractStatefulMcpE2eTest;
+import dev.tachyonmcp.e2e.mcp.McpSdkContract;
 
 class StatefulMcpSdkTest extends AbstractStatefulMcpE2eTest implements McpSdkContract {
