+++ e2e/src/test/java/dev/tachyonmcp/e2e/mcp/v2025_11_25/StatelessServerTest.java
@@ -1,10 +1,11 @@
 /* Copyright (c) 2026 Konstantin Pavlov/IT Staff and contributors. */
-package dev.tachyonmcp.e2e.mcp;
+package dev.tachyonmcp.e2e.mcp.v2025_11_25;
 
 import static net.javacrumbs.jsonunit.assertj.JsonAssertions.assertThatJson;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import dev.tachyonmcp.core.server.TachyonServer;
+import dev.tachyonmcp.e2e.mcp.EchoToolHandler;
 import dev.tachyonmcp.testkit.Mcp20251125Client;
 import java.net.URI;
 import java.net.http.HttpClient;
